# Fix fast 1-weso parameter approximation, for massive speedup