@wjblanke 2024-10-08
merged here because it needed additional unrelated changes for CI but will be sure to credit you
https://github.com/Chia-Network/chiavdf/pull/215